remove interstitia router class
